Move provider fetching to its own file
